requirements.txt
mysql_connector_python==8.0.17
python_dateutil==2.8.1
PrettyTable
